feat: relocated sidebar to shared components
