--- src/components/Header/HeaderDrawer/index.ts
@@ -1,1 +1,0 @@
-export { HeaderDrawer } from './component';
