+++ src/components/SideBar/SideBarList/ListItem/component.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ListItemButton } from '@mui/material';
 
-import MuiLink from '../../../../shared/MuiLink';
+import MuiLink from '../../../shared/MuiLink';
 import { RoutesData } from '../../types';
 
 export function ListItem({ label, link }: RoutesData): JSX.Element {
