+++ src/components/SideBar/index.ts
@@ -1,0 +1,1 @@
+export { SideBar } from './component';
