+++ src/components/layout/Layout/component.tsx
@@ -3,6 +3,7 @@ import { Box, Grid } from '@mui/material';
 import classnames from 'classnames';
 
 import Header from '../../Header';
+import { SideBar } from '../../SideBar';
 
 import { LayoutProps } from './type';
 import { useStyles } from './styles';
@@ -15,6 +16,7 @@ export function Layout({ children }: LayoutProps): JSX.Element {
 
   return (
     <Box>
+      <SideBar isOpen={isSideBar} setOpen={setSideBar} />
       <Header isSideBar={isSideBar} setSideBar={setSideBar} />
       <Grid container spacing={4}>
         <Grid item className={classnames(classes.mainContentWr, { [classes.contentWithSideBar]: isSideBar })}>
